Refactored AppBars
