+++ app/src/main/java/com/example/android/january2022/ui/session/SessionAppBar.kt
@@ -5,12 +5,13 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.MoreVert
-import androidx.compose.material.icons.outlined.Timer
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.example.android.january2022.ui.session.appbaricons.TimerAction
+import com.example.android.january2022.ui.session.actions.ActionSpacer
+import com.example.android.january2022.ui.session.actions.ActionSpacerStart
+import com.example.android.january2022.ui.session.actions.TimerAction
 import com.example.android.january2022.utils.Event
 
 @Composable
@@ -21,11 +22,11 @@ fun SessionAppBar(
 ) {
   BottomAppBar(
     actions = {
-      Spacer(modifier = Modifier.width(4.dp))
+      ActionSpacerStart()
       IconButton(onClick = { /*TODO*/ }) {
         Icon(imageVector = Icons.Default.MoreVert, contentDescription = "Options")
       }
-      Spacer(modifier = Modifier.width(8.dp))
+      ActionSpacer()
       TimerAction(timerVisible,onTimerPress)
     },
     floatingActionButton = {
