+++ app/src/main/java/com/example/android/january2022/ui/session/actions/Spacers.kt
@@ -1,0 +1,17 @@
+package com.example.android.january2022.ui.session.actions
+
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.width
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun ActionSpacer() {
+  Spacer(modifier = Modifier.width(8.dp))
+}
+
+@Composable
+fun ActionSpacerStart() {
+  Spacer(modifier = Modifier.width(4.dp))
+}
